third_party/blink/web_tests/fast/forms/file/file-setrangetext-expected.txt

Test setRangeText() method is not available in file inputs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Running tests on input with attributes: {"type":"file"}

PASS element.value = '0123456789XYZ' threw exception InvalidStateError: Failed to set the 'value' property on 'HTMLInputElement': This input element accepts a filename, which may only be programmatically set to the empty string..
PASS element.setRangeText('ABC', 0, 0) threw exception InvalidStateError: Failed to execute 'setRangeText' on 'HTMLInputElement': The input element's type ('file') does not support selection..
PASS element.value is ""
PASS successfullyParsed is true

TEST COMPLETE
